Separate into internal object-store crate
