+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["object-store"]
+members = ["object-store", "object-store-internal"]
