+++ object-store-internal/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "object-store-internal"
+version = "0.1.10"
+edition = "2021"
+authors = ["Robert Pack <robstar.pack@gmail.com>"]
+description = "A generic object store interface for uniformly interacting with AWS S3, Google Cloud Storage, Azure Storage and local files."
+
+[dependencies]
+async-trait = "0.1.57"
+bytes = "1.2.1"
+futures = "0.3"
+once_cell = "1.12.0"
+object_store = { version = "0.5.6", features = [
+  "azure",
+  "aws",
+  "gcp",
+  "aws_profile",
+] }
+percent-encoding = "2"
+pyo3 = { version = "0.18", features = ["abi3", "abi3-py38"] }
+thiserror = "1.0.34"
+tokio = { version = "1.0", features = [
+  "macros",
+  "rt",
+  "rt-multi-thread",
+  "sync",
+] }
+url = "2.3"
+
+# reqwest is pulled in by object store, but not used by python binding itself
+# for binary wheel best practice, statically link openssl
+reqwest = { version = "*", features = ["native-tls-vendored"] }
+
+[lib]
+crate-type = ["rlib"]
